refactor: extra label logic
