+++ packages/article/article-header-label/article-header-label.js
@@ -1,0 +1,5 @@
+import React from "react";
+import { View } from "react-native";
+import BaseLabel from "./article-header-label.base";
+
+export default BaseLabel((props, label) => <View {...props}>{label}</View>);
