+++ packages/article/article-header-label/article-header-label.web.js
@@ -1,0 +1,7 @@
+import React from "react";
+import BaseLabel from "./article-header-label.base";
+import { LabelContainer } from "../styles/article-header/responsive";
+
+export default BaseLabel((props, label) => (
+  <LabelContainer {...props}>{label}</LabelContainer>
+));
